+++ CCSODiscordBot/Modules/UserManagement/RolePrompt/RolePromptEmbeds.cs
